requirements.txt
aiohttp==3.7.4.post0
discord.py==2.1.1
pytz==2022.7.1